Mobile bug fix and loading screen
Fixed bug: Mobile direct debits were not  displayed on mobile. This was due to states was not setting quick enough on mobiles, fix was to add budget id to session storage to ensure budget id could be used in request.

Loading screen added: When react is mounting components to the display, a white screen would show if the connection is slow. a spinner is now displayed so the user is aware the page is loading.